fix migration + database reset

+++ database/migrations/2026_04_23_065733_drop_category_id_from_posts_table.php
@@ -6,22 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-public function up(): void
-{
-    Schema::table('posts', function (Blueprint $table) {
-        $table->dropForeign(['category_id']);
-        $table->dropColumn('category_id');
-    });
-}
+    public function up(): void
+    {
+        Schema::table('posts', function (Blueprint $table) {
+            if (Schema::hasColumn('posts', 'category_id')) {
+                $table->dropColumn('category_id');
+            }
+        });
+    }
 
-public function down(): void
-{
-    Schema::table('posts', function (Blueprint $table) {
-        $table->unsignedBigInteger('category_id')->nullable();
-
-        $table->foreign('category_id')
-              ->references('id')
-              ->on('categories')
-              ->nullOnDelete();
-    });
-}
+    public function down(): void
+    {
+        Schema::table('posts', function (Blueprint $table) {
+            $table->unsignedBigInteger('category_id')->nullable();
+        });
+    }
+};
